requirements.txt
fastapi
uvicorn
openai==0.27.8
python-dotenv==1.0.0
numpy==1.24.3
scikit-learn==1.3.0
pydantic
tiktoken
gspread
oauth2client
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
flask-cors==3.0.10
